move a lot of renderer init into renderdevice
